+++ docs/make.jl
@@ -1,0 +1,6 @@
+using Documenter, Vela
+
+makedocs(
+    sitename = "Vela.jl",
+    # format = Documenter.LaTeX(platform = "none")
+)